refactor: move to smaller methods in service folder
